qloo-main/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
folium>=0.14.0
streamlit-folium>=0.15.0
geopy>=2.4.0
plotly>=5.17.0
boto3>=1.34.0
botocore>=1.34.0
